finish usage and refactor tests
