+++ packages/typespec-client-generator-core/test/decorators.test.ts
@@ -1749,944 +1749,6 @@ describe("typespec-client-generator-core: decorators", () => {
     });
   });
 
-  describe("@access", () => {
-    describe("namespace access override", () => {
-      it("should inherit access from parent namespace", async () => {
-        const { Test } = (await runner.compile(`
-          @access(Access.public)
-          @service({title: "Test Service"}) namespace TestService;
-          @test
-          model Test {
-            prop: string;
-          }
-        `)) as { Test: Operation };
-
-        const actual = getAccess(runner.context, Test);
-        strictEqual(actual, "public");
-      });
-
-      it("should tag anonymous models with default access", async () => {
-        const { Test, prop } = (await runner.compile(`
-          @access(Access.public)
-          @service({title: "Test Service"}) namespace TestService;
-          @test
-          model Test {
-            @test
-            prop: {
-               foo: string;
-            }
-          }
-        `)) as { Test: Operation; prop: ModelProperty };
-
-        const actual = getAccess(runner.context, Test);
-        const actualAnonymous = getAccess(runner.context, prop.type as Model);
-        strictEqual(actual, "public");
-        strictEqual(actualAnonymous, "public");
-      });
-
-      it("should tag as internal anonymous models with default access", async () => {
-        const { Test, prop } = (await runner.compile(`
-          @access(Access.internal)
-          @service({title: "Test Service"}) namespace TestService;
-          @test
-          model Test {
-            @test
-            prop: {
-               foo: string;
-            }
-          }
-        `)) as { Test: Operation; prop: ModelProperty };
-
-        const actual = getAccess(runner.context, Test);
-        const actualAnonymous = getAccess(runner.context, prop.type as Model);
-        strictEqual(actual, "internal");
-        strictEqual(actualAnonymous, "internal");
-      });
-
-      it("should honor the granular override over the namespace one", async () => {
-        const { Test } = (await runner.compile(`
-          @access(Access.public)
-          @service({title: "Test Service"}) namespace TestService;
-          @access(Access.internal)
-          @test
-          model Test {
-            prop: string;
-          }
-        `)) as { Test: Operation };
-
-        const actual = getAccess(runner.context, Test);
-        strictEqual(actual, "internal");
-      });
-
-      it("locally mark an operation as internal", async () => {
-        const { test } = (await runner.compile(`
-          @access(Access.public)
-          @service({title: "Test Service"}) namespace TestService;
-          @test
-          @access(Access.internal)
-          op test(): void;
-        `)) as { test: Operation };
-
-        const actual = getAccess(runner.context, test);
-        strictEqual(actual, "internal");
-      });
-
-      it("locally mark an operation as public", async () => {
-        const { test } = (await runner.compile(`
-          @access(Access.public)
-          @service({title: "Test Service"}) namespace TestService;
-          @test
-          op test(): void;
-        `)) as { test: Operation };
-
-        const actual = getAccess(runner.context, test);
-        strictEqual(actual, "public");
-      });
-
-      it("mark an operation as internal through the namespace", async () => {
-        const { test } = (await runner.compile(`
-          @access(Access.internal)
-          @service({title: "Test Service"}) namespace TestService;
-          @test
-          op test(): void;
-        `)) as { test: Operation };
-
-        const actual = getAccess(runner.context, test);
-        strictEqual(actual, "internal");
-      });
-    });
-
-    it("default calculated value of operation is undefined, default value of calculated model is undefined", async () => {
-      const { test, Test } = (await runner.compile(`
-        @test
-        model Test{}
-
-        @test
-        op test(): void;
-      `)) as { test: Operation; Test: Model };
-
-      strictEqual(getAccess(runner.context, test), "public");
-      strictEqual(getAccess(runner.context, Test), "public");
-    });
-
-    it("model access calculated by operation", async () => {
-      const { Test, func } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          model Test {
-            prop: string;
-          }
-          @test
-          @access(Access.internal)
-          op func(
-            @body body: Test
-          ): void;
-        }
-      `)) as { Test: Model; func: Operation };
-
-      let actual = getAccess(runner.context, Test);
-      strictEqual(actual, "internal");
-      actual = getAccess(runner.context, func);
-      strictEqual(actual, "internal");
-    });
-
-    it("override calculated model with public access", async () => {
-      const { Test, func } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          @access(Access.public)
-          model Test {
-            prop: string;
-          }
-          @test
-          @access(Access.internal)
-          op func(
-            @body body: Test
-          ): void;
-        }
-      `)) as { Test: Model; func: Operation };
-
-      let actual = getAccess(runner.context, Test);
-      strictEqual(actual, "public");
-      actual = getAccess(runner.context, func);
-      strictEqual(actual, "internal");
-    });
-
-    it("override calculated model with internal access", async () => {
-      const { Test, func } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          @access(Access.internal) // This is an incorrect usage. We will have linter to ban.
-          model Test {
-            prop: string;
-          }
-          @test
-          op func(
-            @body body: Test
-          ): void;
-        }
-        `)) as { Test: Model; func: Operation };
-
-      strictEqual(getAccess(runner.context, Test), "internal");
-      strictEqual(getAccess(runner.context, func), "public");
-    });
-
-    it("access propagation", async () => {
-      const { Fish, Shark, Salmon, SawShark, Origin } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @discriminator("kind")
-          @test
-          model Fish {
-            age: int32;
-          }
-
-          @discriminator("sharktype")
-          @test
-          model Shark extends Fish {
-            kind: "shark";
-            origin: Origin;
-          }
-
-          @test
-          model Salmon extends Fish {
-            kind: "salmon";
-          }
-
-          @test
-          model SawShark extends Shark {
-            sharktype: "saw";
-          }
-
-          @test
-          model Origin {
-            country: string;
-            city: string;
-            manufacture: string;
-          }
-
-          @get
-          @access(Access.internal)
-          op getModel(): Fish;
-        }
-      `)) as { Fish: Model; Shark: Model; Salmon: Model; SawShark: Model; Origin: Model };
-
-      let actual = getAccess(runner.context, Fish);
-      strictEqual(actual, "internal");
-      actual = getAccess(runner.context, Shark);
-      strictEqual(actual, "internal");
-      actual = getAccess(runner.context, Salmon);
-      strictEqual(actual, "internal");
-      actual = getAccess(runner.context, SawShark);
-      strictEqual(actual, "internal");
-      actual = getAccess(runner.context, Origin);
-      strictEqual(actual, "internal");
-    });
-
-    it("complicated access propagation", async () => {
-      const { Test1, Test2, Test3, Test4, Test5, Test6, func1, func2, func3, func4, func5 } =
-        (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          model Test1 {
-            prop: Test2;
-          }
-          @test
-          model Test2 {
-            prop: string;
-          }
-          @test
-          @access(Access.internal)
-          @route("/func1")
-          op func1(
-            @body body: Test1
-          ): void;
-
-          @test
-          model Test3 {
-            prop: string;
-          }
-          @test
-          @access(Access.internal)
-          @route("/func2")
-          op func2(
-            @body body: Test3
-          ): void;
-          @test
-          @route("/func3")
-          op func3(
-            @body body: Test3
-          ): void;
-
-          @test
-          model Test4 {
-            prop: Test5;
-          }
-          @test
-          model Test5 {
-            prop: Test6;
-          }
-          @test
-          model Test6 {
-            prop: string;
-          }
-          @test
-          @access(Access.internal)
-          @route("/func4")
-          op func4(
-            @body body: Test4
-          ): void;
-          @test
-          @route("/func5")
-          op func5(
-            @body body: Test6
-          ): void;
-        }
-      `)) as {
-          Test1: Model;
-          Test2: Model;
-          Test3: Model;
-          Test4: Model;
-          Test5: Model;
-          Test6: Model;
-          func1: Operation;
-          func2: Operation;
-          func3: Operation;
-          func4: Operation;
-          func5: Operation;
-        };
-
-      strictEqual(getAccess(runner.context, func1), "internal");
-      strictEqual(getAccess(runner.context, func2), "internal");
-      strictEqual(getAccess(runner.context, func3), "public");
-      strictEqual(getAccess(runner.context, func4), "internal");
-      strictEqual(getAccess(runner.context, func5), "public");
-
-      strictEqual(getAccess(runner.context, Test1), "internal");
-      strictEqual(getAccess(runner.context, Test2), "internal");
-      strictEqual(getAccess(runner.context, Test3), "public");
-      strictEqual(getAccess(runner.context, Test4), "internal");
-      strictEqual(getAccess(runner.context, Test5), "internal");
-      strictEqual(getAccess(runner.context, Test6), "public");
-    });
-
-    it("access propagation for properties, base models and sub models", async () => {
-      const {
-        Fish,
-        Salmon,
-        Origin,
-        BaseModel,
-        ModelA,
-        ModelB,
-        ModelC,
-        ModelD,
-        ModelE,
-        ModelF,
-        EnumA,
-        func1,
-        func2,
-        func3,
-        func4,
-      } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @discriminator("kind")
-          @test
-          model Fish {
-            age: int32;
-          }
-
-          @test
-          model Origin {
-            country: string;
-            city: string;
-            manufacture: string;
-          }
-
-          @test
-          model Salmon extends Fish {
-            kind: "salmon";
-            origin: Origin;
-          }
-
-          @test
-          model BaseModel {
-            base: string;
-          }
-
-          @test
-          model ModelA extends BaseModel {
-            prop1: ModelB;
-            prop2: ModelC[];
-            prop3: Record<ModelD>;
-            prop4: EnumA;
-            prop5: ModelE | ModelF;
-          }
-
-          @test
-          model ModelB {
-            prop: string;
-          }
-
-          @test
-          model ModelC {
-            prop: string;
-          }
-
-          @test
-          model ModelD {
-            prop: string;
-          }
-
-          @test
-          model ModelE {
-            prop: string;
-          }
-
-          @test
-          model ModelF {
-            prop: string;
-          }
-
-          @test
-          enum EnumA {
-            one,
-            two,
-            three,
-          }
-
-          @test
-          @access(Access.internal)
-          @route("/func1")
-          op func1(
-            @body body: Fish
-          ): void;
-          @test
-          @route("/func2")
-          op func2(
-            @body body: Fish
-          ): void;
-
-          @test
-          @access(Access.internal)
-          @route("/func3")
-          op func3(
-            @body body: ModelA
-          ): void;
-          @test
-          @route("/func4")
-          op func4(
-            @body body: ModelA
-          ): void;
-        }
-      `)) as {
-        Fish: Model;
-        Salmon: Model;
-        Origin: Model;
-        BaseModel: Model;
-        ModelA: Model;
-        ModelB: Model;
-        ModelC: Model;
-        ModelD: Model;
-        ModelE: Model;
-        ModelF: Model;
-        EnumA: Model;
-        func1: Operation;
-        func2: Operation;
-        func3: Operation;
-        func4: Operation;
-      };
-
-      strictEqual(getAccess(runner.context, func1), "internal");
-      strictEqual(getAccess(runner.context, func2), "public");
-      strictEqual(getAccess(runner.context, func3), "internal");
-      strictEqual(getAccess(runner.context, func4), "public");
-
-      strictEqual(getAccess(runner.context, Fish), "public");
-      strictEqual(getAccess(runner.context, Salmon), "public");
-      strictEqual(getAccess(runner.context, Origin), "public");
-      strictEqual(getAccess(runner.context, BaseModel), "public");
-      strictEqual(getAccess(runner.context, ModelA), "public");
-      strictEqual(getAccess(runner.context, ModelB), "public");
-      strictEqual(getAccess(runner.context, ModelC), "public");
-      strictEqual(getAccess(runner.context, ModelD), "public");
-      strictEqual(getAccess(runner.context, ModelE), "public");
-      strictEqual(getAccess(runner.context, ModelF), "public");
-      strictEqual(getAccess(runner.context, EnumA), "public");
-    });
-
-    it("access propagation with override", async () => {
-      const {
-        Test1,
-        Test2,
-        Test3,
-        Test4,
-        Test5,
-        func1,
-        func2,
-        func3,
-        func4,
-        func5,
-        func6,
-        func7,
-        func8,
-      } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          model Test1 {
-          }
-          @test
-          @access(Access.internal)
-          @route("/func1")
-          op func1(
-            @body body: Test1
-          ): void;
-
-          @test
-          model Test2 {
-          }
-          @test
-          @route("/func2")
-          op func2(
-            @body body: Test2
-          ): void;
-
-          @test
-          model Test3 {
-          }
-          @test
-          @access(Access.public)
-          @route("/func3")
-          op func3(
-            @body body: Test3
-          ): void;
-
-
-          @test
-          model Test4 {
-          }
-          @test
-          @access(Access.internal)
-          @route("/func4")
-          op func4(
-            @body body: Test4
-          ): void;
-          @test
-          @route("/func5")
-          op func5(
-            @body body: Test4
-          ): void;
-
-          @test
-          model Test5 {
-          }
-          @test
-          @access(Access.internal)
-          @route("/func6")
-          op func6(
-            @body body: Test5
-          ): void;
-          @test
-          @route("/func7")
-          op func7(
-            @body body: Test5
-          ): void;
-          @test
-          @access(Access.public)
-          @route("/func8")
-          op func8(
-            @body body: Test5
-          ): void;
-        }
-      `)) as {
-        Test1: Model;
-        Test2: Model;
-        Test3: Model;
-        Test4: Model;
-        Test5: Model;
-        func1: Operation;
-        func2: Operation;
-        func3: Operation;
-        func4: Operation;
-        func5: Operation;
-        func6: Operation;
-        func7: Operation;
-        func8: Operation;
-      };
-
-      strictEqual(getAccess(runner.context, func1), "internal");
-      strictEqual(getAccess(runner.context, func2), "public");
-      strictEqual(getAccess(runner.context, func3), "public");
-      strictEqual(getAccess(runner.context, func4), "internal");
-      strictEqual(getAccess(runner.context, func5), "public");
-      strictEqual(getAccess(runner.context, func6), "internal");
-      strictEqual(getAccess(runner.context, func7), "public");
-      strictEqual(getAccess(runner.context, func8), "public");
-
-      strictEqual(getAccess(runner.context, Test1), "internal");
-      strictEqual(getAccess(runner.context, Test2), "public");
-      strictEqual(getAccess(runner.context, Test3), "public");
-      strictEqual(getAccess(runner.context, Test4), "public");
-      strictEqual(getAccess(runner.context, Test5), "public");
-    });
-
-    it("access propagation with nullable", async () => {
-      await runner.compileWithBuiltInService(
-        `
-        model RunStep {
-          id: string;
-          lastError: RunStepError | null;
-        }
-
-        model RunStepError {
-          code: string;
-          message: string;
-        }
-
-        @get
-        @route("/threads/{threadId}/runs/{runId}/steps/{stepId}")
-        op getRunStep(
-          @path threadId: string,
-          @path runId: string,
-          @path stepId: string,
-        ): RunStep;
-
-        @get
-        @route("/threads/{threadId}/runs/{runId}/steps")
-        op listRunSteps(
-          @path threadId: string,
-          @path runId: string,
-        ): RunStep[];
-        @@access(listRunSteps, Access.internal);
-        `
-      );
-      const models = runner.context.sdkPackage.models;
-      strictEqual(models.length, 2);
-      strictEqual(models[0].access, "public");
-      strictEqual(models[1].access, "public");
-    });
-  });
-
-  describe("@usage", () => {
-    it("defaults calculated usage", async () => {
-      const { Model1, Model2, Model3, Model4 } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          model Model1{ prop: string }
-
-          @test
-          model Model2{ prop: string }
-
-          @test
-          model Model3{ prop: string }
-
-          @test
-          model Model4 { prop: string }
-
-          @test
-          @route("/func1")
-          op func1(@body body: Model1): void;
-
-          @test
-          @route("/func2")
-          op func2(): Model2;
-
-          @test
-          @route("/func3")
-          op func3(@body body: Model3): Model3;
-        }
-      `)) as { Model1: Model; Model2: Model; Model3: Model; Model4: Model };
-
-      strictEqual(getUsage(runner.context, Model1), UsageFlags.Input | UsageFlags.Json);
-      strictEqual(getUsage(runner.context, Model2), UsageFlags.Output | UsageFlags.Json);
-      strictEqual(
-        getUsage(runner.context, Model3),
-        UsageFlags.Input | UsageFlags.Output | UsageFlags.Json
-      );
-      strictEqual(getUsage(runner.context, Model4), UsageFlags.None);
-    });
-
-    it("usage override", async () => {
-      const { Model1, Model2, Model3, Model4, Enum1, Enum2 } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          @usage(Usage.input | Usage.output)
-          enum Enum1{
-            one,
-            two,
-            three
-          }
-
-          @test
-          enum Enum2{
-            one,
-            two,
-            three
-          }
-
-          @test
-          @usage(Usage.input | Usage.output)
-          model Model1{ prop: string }
-
-          @test
-          model Model4{ prop: string }
-
-          @test
-          @usage(Usage.output)
-          model Model2{ prop: string }
-
-          @test
-          @usage(Usage.input)
-          model Model3{ prop: string }
-
-          @test
-          @route("/func1")
-          op func1(@body body: Model2): void;
-
-          @test
-          @route("/func2")
-          op func2(): Model3;
-        }
-      `)) as {
-        Model1: Model;
-        Model2: Model;
-        Model3: Model;
-        Model4: Model;
-        Enum1: Enum;
-        Enum2: Enum;
-      };
-
-      strictEqual(getUsage(runner.context, Model1), UsageFlags.Input | UsageFlags.Output);
-      strictEqual(
-        getUsage(runner.context, Model2),
-        UsageFlags.Input | UsageFlags.Output | UsageFlags.Json
-      );
-      strictEqual(
-        getUsage(runner.context, Model3),
-        UsageFlags.Input | UsageFlags.Output | UsageFlags.Json
-      );
-      strictEqual(getUsage(runner.context, Model4), UsageFlags.None);
-      strictEqual(getUsage(runner.context, Enum1), UsageFlags.Input | UsageFlags.Output);
-      strictEqual(getUsage(runner.context, Enum2), UsageFlags.None);
-    });
-
-    it("wrong usage value", async () => {
-      const diagnostics = await runner.diagnose(`
-        @test
-        @usage(1)
-        model Model1{}
-      `);
-
-      expectDiagnostics(diagnostics, {
-        code: "invalid-argument",
-      });
-    });
-
-    it("usage propagation", async () => {
-      const { Fish, Shark, Salmon, SawShark, Origin } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @discriminator("kind")
-          @test
-          model Fish {
-            age: int32;
-          }
-
-          @discriminator("sharktype")
-          @test
-          @usage(Usage.input)
-          model Shark extends Fish {
-            kind: "shark";
-            origin: Origin;
-          }
-
-          @test
-          model Salmon extends Fish {
-            kind: "salmon";
-          }
-
-          @test
-          model SawShark extends Shark {
-            sharktype: "saw";
-          }
-
-          @test
-          model Origin {
-            country: string;
-            city: string;
-            manufacture: string;
-          }
-
-          @get
-          op getModel(): Fish;
-        }
-      `)) as { Fish: Model; Shark: Model; Salmon: Model; SawShark: Model; Origin: Model };
-
-      strictEqual(getUsage(runner.context, Fish), UsageFlags.Output | UsageFlags.Json);
-      strictEqual(
-        getUsage(runner.context, Shark),
-        UsageFlags.Input | UsageFlags.Output | UsageFlags.Json
-      );
-      strictEqual(getUsage(runner.context, Salmon), UsageFlags.Output | UsageFlags.Json);
-      strictEqual(getUsage(runner.context, SawShark), UsageFlags.Output | UsageFlags.Json);
-      strictEqual(getUsage(runner.context, Origin), UsageFlags.Output | UsageFlags.Json);
-    });
-
-    it("usage and convenience", async () => {
-      const { Fish } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          model Fish {
-            age: int32;
-          }
-
-          @put
-          @convenientAPI(true)
-          op putModel(@body body: Fish): void;
-
-          @get
-          @convenientAPI(false)
-          op getModel(): Fish;
-        }
-      `)) as { Fish: Model };
-
-      strictEqual(getUsage(runner.context, Fish), UsageFlags.Input | UsageFlags.Json);
-
-      const { Dog } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          model Dog {
-            age: int32;
-          }
-
-          @put
-          @convenientAPI(false)
-          op putModel(@body body: Dog): void;
-
-          @get
-          @convenientAPI(true)
-          op getModel(): Dog;
-        }
-      `)) as { Dog: Model };
-
-      strictEqual(getUsage(runner.context, Dog), UsageFlags.Output | UsageFlags.Json);
-    });
-
-    it("patch usage", async () => {
-      const { PatchModel, JsonMergePatchModel } = (await runner.compile(`
-        @service({})
-        @test namespace MyService {
-          @test
-          model PatchModel {
-            age: int32;
-          }
-
-          @test
-          model JsonMergePatchModel {
-            prop: string
-          }
-
-          @patch
-          @route("/patch")
-          op patchModel(@body body: PatchModel): void;
-
-          @patch
-          @route("/jsonMergePatch")
-          op jsonMergePatchModel(@body body: JsonMergePatchModel, @header contentType: "application/merge-patch+json"): void;
-        }
-      `)) as { PatchModel: Model; JsonMergePatchModel: Model };
-
-      strictEqual(getUsage(runner.context, PatchModel), UsageFlags.Input | UsageFlags.Json);
-      strictEqual(
-        getUsage(runner.context, JsonMergePatchModel),
-        UsageFlags.JsonMergePatch | UsageFlags.Input | UsageFlags.Json
-      );
-    });
-
-    it("@usage Input and Output on Namespace", async () => {
-      const { OrphanModel, InputModel, OutputModel, RoundtripModel } = (await runner.compile(`
-        @service({})
-        @test
-        @usage(Usage.input | Usage.output)
-        namespace MyService {
-          @test
-          model OrphanModel {
-            prop: string;
-          }
-
-          @test
-          model InputModel {
-            prop: string
-          }
-
-          @test
-          model OutputModel {
-            prop: string
-          }
-
-          @test
-          model RoundtripModel {
-            prop: string
-          }
-
-          @route("/one")
-          op one(@body body: InputModel): OutputModel;
-
-          @route("/two")
-          op two(@body body: RoundtripModel): RoundtripModel;
-        }
-      `)) as { OrphanModel: Model; InputModel: Model; OutputModel: Model; RoundtripModel: Model };
-      strictEqual(getUsage(runner.context, OrphanModel), UsageFlags.Input | UsageFlags.Output);
-      // this is set to input and output because of the namespace override
-      strictEqual(
-        getUsage(runner.context, InputModel),
-        UsageFlags.Input | UsageFlags.Output | UsageFlags.Json
-      );
-      strictEqual(
-        getUsage(runner.context, OutputModel),
-        UsageFlags.Input | UsageFlags.Output | UsageFlags.Json
-      );
-      strictEqual(
-        getUsage(runner.context, RoundtripModel),
-        UsageFlags.Input | UsageFlags.Output | UsageFlags.Json
-      );
-    });
-
-    it("@usage namespace override", async () => {
-      const { OrphanModel, OrphanModelWithOverride } = (await runner.compile(`
-        @service({})
-        @test
-        @usage(Usage.input)
-        namespace MyService {
-          @test
-          model OrphanModel {
-            prop: string;
-          }
-
-          @test
-          @usage(Usage.input | Usage.output)
-          model OrphanModelWithOverride {
-            prop: string;
-          }
-        }
-      `)) as { OrphanModel: Model; OrphanModelWithOverride: Model };
-      strictEqual(getUsage(runner.context, OrphanModel), UsageFlags.Input);
-      strictEqual(
-        getUsage(runner.context, OrphanModelWithOverride),
-        UsageFlags.Input | UsageFlags.Output
-      );
-    });
-  });
-
   describe("@flattenProperty", () => {
     it("marks a model property to be flattened with suppression of deprecation warning", async () => {
       await runner.compileWithBuiltInService(`
